Fix issue #1 which makes negative duration possible
https://github.com/poshjosh/rate-limiter/issues/1.

If a bandwidth is created at time t micros, that bandwidth
may only be updated at times later than t micros.

We also added documentation to the io.github.poshjosh.ratelimiter.util.Ticker
class advising the user of an instance of Ticker to keep track of time
with respect to bandwidths.